requirements.txt
matplotlib == 3.9.2
pillow == 11.0.0
torch == 2.5.0
transformers == 4.46.0
numpy == 2.1.2
open3d == 0.18.0 
opencv-python-headless == 4.10.0.84
plotly == 5.24.1
streamlit == 1.39.0 
